requirements.txt
backcall==0.1.0
colorama==0.4.1
decorator==4.3.0
Django==1.11.11
django-redis<=3.8.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
parso==0.3.1
pickleshare==0.7.5
prompt-toolkit==2.0.7
Pygments==2.3.1
pytz==2018.7
redis==3.0.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
win-unicode-console==0.5
PyMySQL==0.9.3
djangorestframework==3.8.2
coreapi==1.32.0
markdown==2.1.0
django-filter==1.0.1
django-crispy-forms==1.7.2
django-guardian==1.1.1
